requirements.txt
python-dotenv
gradio
elevenlabs
openai

# streamlit
# transformers[torch]
# torchaudio